backend/app/services/rag.py
"""
AIEco - RAG Service
Document ingestion, vector search, and retrieval
"""

from typing import Any, Dict, List, Optional
import uuid
import hashlib
from datetime import datetime
import structlog

logger = structlog.get_logger()


class RAGService:
    """
    RAG (Retrieval Augmented Generation) Service
    Handles document ingestion, embedding, and semantic search
    """
    
    def __init__(self, chroma_client=None, embedding_client=None):
        self.chroma_client = chroma_client
        self.embedding_client = embedding_client
        self._collections: Dict[str, Any] = {}
    
    async def ingest_document(
        self,
        content: bytes,
        filename: str,
        collection: str = "default",
        user_id: str = None,
        chunk_size: int = 1000,
        chunk_overlap: int = 200
    ) -> Dict[str, Any]:
        """
        Ingest a document into the vector store.
        
        1. Parse document based on file type
        2. Split into chunks
        3. Generate embeddings
        4. Store in vector database
        """
        document_id = str(uuid.uuid4())
        
        logger.info(
            "📄 Ingesting document",
            document_id=document_id,
            filename=filename,
            collection=collection
        )
        
        # Parse document
        text = await self._parse_document(content, filename)
        
        # Split into chunks
        chunks = self._split_text(text, chunk_size, chunk_overlap)
        
        # Generate embeddings and store
        chunk_ids = []
        for i, chunk in enumerate(chunks):
            chunk_id = f"{document_id}_{i}"
            
            # Store in vector DB (simplified - real impl would batch)
            await self._store_chunk(
                chunk_id=chunk_id,
                content=chunk,
                collection=collection,
                metadata={
                    "document_id": document_id,
                    "filename": filename,
                    "chunk_index": i,
                    "user_id": user_id,
                    "created_at": datetime.utcnow().isoformat()
                }
            )
            chunk_ids.append(chunk_id)
        
        logger.info(
            "✅ Document ingested",
            document_id=document_id,
            chunks=len(chunks)
        )
        
        return {
            "document_id": document_id,
            "filename": filename,
            "chunk_count": len(chunks),
            "chunk_ids": chunk_ids
        }
    
    async def search(
        self,
        query: str,
        collection: str = "default",
        top_k: int = 5,
        filters: Optional[Dict] = None
    ) -> List[Dict[str, Any]]:
        """
        Semantic search for relevant document chunks.
        """
        logger.debug("🔍 RAG search", query=query[:50], collection=collection)
        
        # For demo - return mock results
        # Real implementation would:
        # 1. Generate query embedding
        # 2. Search vector DB
        # 3. Return ranked results
        
        results = [
            {
                "id": "chunk_1",
                "content": f"Relevant content for: {query}",
                "score": 0.95,
                "metadata": {
                    "source": "documentation.md",
                    "page": 1
                }
            }
        ]
        
        return results
    
    async def list_collections(self) -> List[Dict[str, Any]]:
        """List all document collections"""
        # Mock implementation
        return [
            {
                "name": "default",
                "document_count": 0,
                "chunk_count": 0,
                "created_at": datetime.utcnow().isoformat()
            }
        ]
    
    async def delete_document(self, document_id: str) -> bool:
        """Delete a document and its chunks from the store"""
        logger.info("🗑️ Deleting document", document_id=document_id)
        # Implementation would delete from vector DB
        return True
    
    async def _parse_document(self, content: bytes, filename: str) -> str:
        """Parse document content based on file type"""
        ext = filename.split(".")[-1].lower() if "." in filename else ""
        
        if ext == "pdf":
            # Use pypdf for PDF parsing
            from pypdf import PdfReader
            from io import BytesIO
            reader = PdfReader(BytesIO(content))
            text = "\n".join(page.extract_text() or "" for page in reader.pages)
            
        elif ext == "docx":
            # Use python-docx for Word documents
            from docx import Document
            from io import BytesIO
            doc = Document(BytesIO(content))
            text = "\n".join(para.text for para in doc.paragraphs)
            
        elif ext in ["txt", "md", "py", "js", "ts", "json", "yaml", "yml"]:
            # Plain text files
            text = content.decode("utf-8", errors="ignore")
            
        else:
            # Try as plain text
            text = content.decode("utf-8", errors="ignore")
        
        return text
    
    def _split_text(
        self,
        text: str,
        chunk_size: int = 1000,
        chunk_overlap: int = 200
    ) -> List[str]:
        """Split text into overlapping chunks"""
        chunks = []
        start = 0
        
        while start < len(text):
            end = start + chunk_size
            chunk = text[start:end]
            
            # Try to break at sentence or paragraph boundary
            if end < len(text):
                # Look for last period, newline, or sentence end
                for sep in ["\n\n", "\n", ". ", "! ", "? "]:
                    last_sep = chunk.rfind(sep)
                    if last_sep > chunk_size // 2:
                        chunk = chunk[:last_sep + len(sep)]
                        break
            
            chunks.append(chunk.strip())
            start = start + len(chunk) - chunk_overlap
        
        return [c for c in chunks if c]  # Filter empty chunks
    
    async def _store_chunk(
        self,
        chunk_id: str,
        content: str,
        collection: str,
        metadata: Dict
    ) -> None:
        """Store a chunk in the vector database"""
        # Real implementation would use ChromaDB
        # self.chroma_client.get_or_create_collection(collection).add(...)
        pass


# Dependency injection
_rag_service: Optional[RAGService] = None


def get_rag_service() -> RAGService:
    """Get or create RAG service instance"""
    global _rag_service
    if _rag_service is None:
        _rag_service = RAGService()
    return _rag_service
