backend/requirements.txt
# AIEco Backend Dependencies
# Python 3.11+

# ===== Core Framework =====
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9

# ===== Authentication =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ===== Database =====
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1

# ===== AI/ML =====
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
langgraph==0.0.26
chromadb==0.4.22
tiktoken==0.5.2

# ===== Document Processing =====
pypdf==4.0.1
python-docx==1.1.0
markdown==3.5.2
beautifulsoup4==4.12.3

# ===== Observability =====
langfuse==2.8.1
prometheus-client==0.19.0
structlog==24.1.0

# ===== Utilities =====
httpx==0.26.0
aiofiles==23.2.1
python-dotenv==1.0.1
tenacity==8.2.3

# ===== Testing =====
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# ===== Development =====
black==24.1.1
ruff==0.2.1
mypy==1.8.0
